Suresh code

+++ global.h
@@ -1,4 +1,4 @@
-/* GLOBAL.H - RSAREF types and constants (rfc1321)
+/* GLOBAL.H - RSAREF types and constants
  */
 
 /* PROTOTYPES should be set to one if and only if the compiler supports
@@ -17,7 +17,7 @@ typedef unsigned char *POINTER;
 typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef unsigned int UINT4;
 
 /* PROTO_LIST is defined depending on how PROTOTYPES is defined above.
 If using PROTOTYPES, then PROTO_LIST returns the list, otherwise it
